requirements.txt
asgiref==3.8.1
chardet==5.2.0
Django==5.1.3
django-cors-headers==4.6.0
django-parler==2.3
django-parler-rest==2.2
djangorestframework==3.15.0
exceptiongroup==1.2.2
flake8==6.1.0
iniconfig==2.0.0
mccabe==0.7.0
packaging==24.1
pillow==11.0.0
pluggy==1.5.0
pycodestyle==2.11.1
pyflakes==3.1.0
pytest==7.4.2
reportlab==4.2.5
sqlparse==0.5.1
tomli==2.0.2
typing_extensions==4.12.2
